Reorganize API into versioned packages
