+++ recipe-server/normandy/recipes/api/v1/serializers.py
