+++ recipe-server/normandy/recipes/tests/api/v1/test_api.py
